Support declaring TS workflows in the feature file
